Rename: 패키지명 변경 (facade -> client)
